+++ two/src/main/java/com/wizardlybump17/physics/two/task/factory/RegisteredTaskFactory.java
@@ -5,15 +5,15 @@ import org.jetbrains.annotations.NotNull;
 
 public class RegisteredTaskFactory {
 
-    public @NotNull RegisteredTask create(long id, @NotNull Runnable task, long delay, long period, long startedAt) {
+    public @NotNull RegisteredTask create(int id, @NotNull Runnable task, long delay, long period, long startedAt) {
         return new RegisteredTask(id, startedAt, delay, period, task);
     }
 
-    public @NotNull RegisteredTask create(long id, @NotNull Runnable task, long startedAt) {
+    public @NotNull RegisteredTask create(int id, @NotNull Runnable task, long startedAt) {
         return create(id, task, 0, 0, startedAt);
     }
 
-    public @NotNull RegisteredTask create(long id, @NotNull Runnable task, long delay, long startedAt) {
+    public @NotNull RegisteredTask create(int id, @NotNull Runnable task, long delay, long startedAt) {
         return create(id, task, delay, 0, startedAt);
     }
 }
